extract neighbor table/info to its own module
